Packaging tweak for RPMS

+++ build-src/before_remove.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+# On RPM update don't remove anything
+if [ "$1" -gt 0 ]; then
+    exit 0
+fi
+
 # Stop, disable and remove services
 systemctl disable tccd tccd-sleep || true
 systemctl stop tccd || true
@@ -7,9 +12,12 @@ rm /etc/systemd/system/tccd.service || true
 rm /etc/systemd/system/tccd-sleep.service || true
 systemctl daemon-reload || true
 
-# Remove log files
-rm -rf /var/log/tcc/ || true
-rm -rf /var/log/tccd/ || true
+# Remove log and config files (unless deb upgrade)
+if [ "$1" -ne "upgrade" ]; then
+    rm -rf /var/log/tcc/ || true
+    rm -rf /var/log/tccd/ || true
+    rm -rf /etc/tcc/ || true
+fi
 
 # Remove link to GUI
 rm -rf /usr/bin/tuxedo-control-center || true
